renamed DateOnlyComparer
