+++ Compare-NET-Objects/RootComparerFactory.cs
@@ -36,7 +36,7 @@ namespace KellermanSoftware.CompareNetObjects
             _rootComparer.TypeComparers = new List<BaseTypeComparer>();
 
             _rootComparer.TypeComparers.Add(new StringComparer(_rootComparer));
-            _rootComparer.TypeComparers.Add(new DateComparer(_rootComparer));
+            _rootComparer.TypeComparers.Add(new DateTimeComparer(_rootComparer));
             _rootComparer.TypeComparers.Add(new DecimalComparer(_rootComparer));
             _rootComparer.TypeComparers.Add(new DoubleComparer(_rootComparer));
             _rootComparer.TypeComparers.Add(new PointerComparer(_rootComparer));
